+++ src/Sandbox.CSharp/Modules/ModuleExtensions.cs
@@ -1,9 +1,6 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 using Microsoft.Extensions.DependencyInjection;
 using Sandbox.CSharp.Core;
+using Sandbox.CSharp.Modules.Async;
 using Sandbox.CSharp.Modules.Paginator;
 
 namespace Sandbox.CSharp.Modules;
@@ -20,7 +17,7 @@ public static class ModuleExtensions
     /// <returns>The <see cref="IServiceCollection"/> so that additional calls can be chained.</returns>
     public static IServiceCollection AddModules(this IServiceCollection services) => services
         .AddCoreModules()
+        .AddAsyncModules()
         .AddPaginatorModule()
-        .AddModule<HelloWorld>()
-        .AddModule<AsyncEnumerable>();
+        .AddModule<HelloWorld>();
 }
